split client websocket into its own file
