+++ lute/net/src/client.cpp
@@ -5,19 +5,14 @@
 #include "lute/userdatas.h"
 
 #include "Luau/DenseHash.h"
-#include "Luau/VecDeque.h"
 
 #include "lua.h"
 #include "lualib.h"
 
 #include "curl/curl.h"
-#include "curl/websockets.h"
 
-#include <atomic>
 #include <cstring>
-#include <functional>
 #include <memory>
-#include <mutex>
 #include <string>
 #include <utility>
 #include <vector>
@@ -25,6 +20,7 @@
 namespace net::client
 {
 struct WebSocketHandle;
+int websocket(lua_State* L);
 int ws_send(lua_State* L);
 int ws_close(lua_State* L);
 }
@@ -92,33 +88,11 @@ static void initializeNetClient(lua_State* L)
 
     lua_setuserdatametatable(L, kWebSocketHandleTag);
 }
-
 } // namespace
 
 namespace net::client
 {
-
 static const std::string kEmptyHeaderKey = "";
-
-struct WebSocketPayload
-{
-    const char* data = nullptr;
-    size_t length = 0;
-    bool binary = false;
-};
-
-struct WebSocketPollState
-{
-    uv_poll_t handle{};
-    std::shared_ptr<WebSocketHandle> owner;
-};
-
-struct PendingSend
-{
-    std::string payload;
-    bool binary = false;
-    size_t offset = 0;
-};
 
 struct CurlResponse
 {
@@ -132,25 +106,6 @@ struct CurlResponse
     {
     }
 };
-
-static WebSocketPayload extractWebSocketPayload(lua_State* L, int index)
-{
-    if (lua_isstring(L, index))
-    {
-        size_t length = 0;
-        const char* data = lua_tolstring(L, index, &length);
-        return {data, length, false};
-    }
-
-    if (lua_isbuffer(L, index))
-    {
-        size_t length = 0;
-        void* data = lua_tobuffer(L, index, &length);
-        return {static_cast<const char*>(data), length, true};
-    }
-
-    luaL_typeerrorL(L, index, "string or buffer");
-}
 
 static size_t writeFunction(char* ptr, size_t size, size_t nmemb, void* userdata)
 {
@@ -245,399 +200,6 @@ static CurlResponse requestData(
     return resp;
 }
 
-struct WebSocketHandle : std::enable_shared_from_this<WebSocketHandle>
-{
-    Runtime* runtime = nullptr;
-    CURL* curl = nullptr;
-    curl_slist* headerList = nullptr;
-    std::shared_ptr<Ref> onOpenRef;
-    std::shared_ptr<Ref> onMessageRef;
-    std::shared_ptr<Ref> onCloseRef;
-    std::shared_ptr<Ref> onErrorRef;
-    WebSocketPollState* pollState = nullptr;
-    std::atomic<bool> isClosed{false};
-    std::atomic<bool> hasScheduledClose{false};
-    std::atomic<bool> hasPendingToken{false};
-    std::mutex curlMutex;
-    int activePollEvents = 0;
-    std::vector<char> recvBuffer = std::vector<char>(16 * 1024);
-    std::string currentMessage;
-    bool currentBinary = false;
-    bool hasCurrentMessage = false;
-    std::string closePayload;
-    Luau::VecDeque<PendingSend> pendingSends;
-
-    void scheduleCallback(const std::shared_ptr<Ref>& callback, std::function<int(lua_State*)> argPusher)
-    {
-        if (!callback || !runtime)
-            return;
-
-        runtime->scheduleLuauCallback(callback, std::move(argPusher));
-    }
-
-    void scheduleCloseCallback(int closeCode = 1000, std::string closeReason = "")
-    {
-        bool expected = false;
-        if (!hasScheduledClose.compare_exchange_strong(expected, true))
-            return;
-
-        scheduleCallback(
-            onCloseRef,
-            [closeCode, closeReason = std::move(closeReason)](lua_State* L)
-            {
-                lua_pushinteger(L, closeCode);
-                lua_pushlstring(L, closeReason.data(), closeReason.size());
-                return 2;
-            }
-        );
-    }
-
-    void stopPolling()
-    {
-        if (!pollState)
-            return;
-
-        WebSocketPollState* state = pollState;
-        pollState = nullptr;
-        activePollEvents = 0;
-
-        uv_poll_stop(&state->handle);
-        uv_close(
-            reinterpret_cast<uv_handle_t*>(&state->handle),
-            [](uv_handle_t* handle)
-            {
-                delete static_cast<WebSocketPollState*>(handle->data);
-            }
-        );
-    }
-
-    void clearPendingSends()
-    {
-        pendingSends.clear();
-    }
-
-    void finishClose(bool sendCloseFrame)
-    {
-        stopPolling();
-        clearPendingSends();
-
-        {
-            std::lock_guard<std::mutex> lock(curlMutex);
-            if (curl)
-            {
-                if (sendCloseFrame)
-                {
-                    size_t sent = 0;
-                    (void)curl_ws_send(curl, "", 0, &sent, 0, CURLWS_CLOSE);
-                }
-
-                curl_easy_cleanup(curl);
-                curl = nullptr;
-            }
-
-            if (headerList)
-            {
-                curl_slist_free_all(headerList);
-                headerList = nullptr;
-            }
-        }
-
-        if (runtime && hasPendingToken.exchange(false))
-            runtime->releasePendingToken();
-    }
-
-    bool applyPollingEvents(int events)
-    {
-        if (!pollState)
-            return false;
-
-        if (activePollEvents == events)
-            return true;
-
-        int startResult = uv_poll_start(
-            &pollState->handle,
-            events,
-            [](uv_poll_t* handle, int status, int events)
-            {
-                auto* state = static_cast<WebSocketPollState*>(handle->data);
-                if (!state)
-                    return;
-
-                if (status < 0)
-                {
-                    state->owner->closeWithError(uv_strerror(status));
-                    return;
-                }
-
-                if (events & UV_READABLE)
-                    state->owner->processIncoming();
-
-                if (!state->owner->isClosed.load() && (events & UV_WRITABLE))
-                    state->owner->flushOutgoing();
-            }
-        );
-
-        if (startResult != 0)
-            return false;
-
-        activePollEvents = events;
-        return true;
-    }
-
-    bool updatePollingInterest()
-    {
-        int events = UV_READABLE;
-        if (!pendingSends.empty())
-            events |= UV_WRITABLE;
-
-        return applyPollingEvents(events);
-    }
-
-    void closeWithCode(int closeCode = 1000, std::string closeReason = "", bool sendCloseFrame = true)
-    {
-        bool expected = false;
-        if (!isClosed.compare_exchange_strong(expected, true))
-            return;
-
-        scheduleCloseCallback(closeCode, std::move(closeReason));
-        finishClose(sendCloseFrame);
-    }
-
-    void closeWithError(std::string error)
-    {
-        bool expected = false;
-        if (!isClosed.compare_exchange_strong(expected, true))
-            return;
-
-        scheduleCallback(
-            onErrorRef,
-            [error = std::move(error)](lua_State* L)
-            {
-                lua_pushlstring(L, error.data(), error.size());
-                return 1;
-            }
-        );
-
-        scheduleCloseCallback(1006);
-        finishClose(false);
-    }
-
-    void processIncoming()
-    {
-        while (!isClosed.load())
-        {
-            size_t receivedLength = 0;
-            const curl_ws_frame* meta = nullptr;
-            CURLcode result = CURLE_OK;
-
-            {
-                std::lock_guard<std::mutex> lock(curlMutex);
-                if (!curl)
-                    return;
-
-                result = curl_ws_recv(curl, recvBuffer.data(), recvBuffer.size(), &receivedLength, &meta);
-            }
-
-            if (isClosed.load())
-                return;
-
-            if (result == CURLE_AGAIN)
-                return;
-
-            if (result != CURLE_OK)
-            {
-                if (isClosed.load() || hasScheduledClose.load())
-                    return;
-
-                closeWithError(curl_easy_strerror(result));
-                return;
-            }
-
-            if (!meta)
-                continue;
-
-            if (meta->flags & CURLWS_CLOSE)
-            {
-                closePayload.append(recvBuffer.data(), receivedLength);
-
-                if (meta->bytesleft != 0)
-                    continue;
-
-                int closeCode = 1000;
-                std::string closeReason;
-
-                if (closePayload.size() >= 2)
-                {
-                    const unsigned char* payload = reinterpret_cast<const unsigned char*>(closePayload.data());
-                    closeCode = int((payload[0] << 8) | payload[1]);
-                    if (closePayload.size() > 2)
-                        closeReason.assign(closePayload.data() + 2, closePayload.size() - 2);
-                }
-
-                closeWithCode(closeCode, std::move(closeReason));
-                return;
-            }
-
-            if (meta->flags & CURLWS_PING)
-            {
-                size_t sent = 0;
-                std::lock_guard<std::mutex> lock(curlMutex);
-                if (curl)
-                    (void)curl_ws_send(curl, recvBuffer.data(), receivedLength, &sent, 0, CURLWS_PONG);
-                continue;
-            }
-
-            if (meta->flags & (CURLWS_TEXT | CURLWS_BINARY | CURLWS_CONT))
-            {
-                if (meta->flags & (CURLWS_TEXT | CURLWS_BINARY))
-                {
-                    bool binary = (meta->flags & CURLWS_BINARY) != 0;
-                    if (!hasCurrentMessage)
-                    {
-                        currentMessage.clear();
-                        currentBinary = binary;
-                        hasCurrentMessage = true;
-                    }
-                    else if (currentBinary != binary)
-                    {
-                        closeWithError("websocket received mixed message types");
-                        return;
-                    }
-                }
-                else if (!hasCurrentMessage)
-                {
-                    currentMessage.clear();
-                    currentBinary = false;
-                    hasCurrentMessage = true;
-                }
-
-                currentMessage.append(recvBuffer.data(), receivedLength);
-
-                if (meta->bytesleft == 0 && !(meta->flags & CURLWS_CONT))
-                {
-                    std::string message = std::move(currentMessage);
-                    bool binary = currentBinary;
-                    hasCurrentMessage = false;
-
-                    scheduleCallback(
-                        onMessageRef,
-                        [message = std::move(message), binary](lua_State* L)
-                        {
-                            if (binary)
-                            {
-                                void* buf = lua_newbuffer(L, message.size());
-                                if (!message.empty())
-                                    memcpy(buf, message.data(), message.size());
-                            }
-                            else
-                            {
-                                lua_pushlstring(L, message.data(), message.size());
-                            }
-                            return 1;
-                        }
-                    );
-                }
-            }
-        }
-    }
-
-    bool startPolling(curl_socket_t socket)
-    {
-        if (!runtime || socket == CURL_SOCKET_BAD)
-            return false;
-
-        auto* state = new WebSocketPollState();
-        state->owner = shared_from_this();
-        state->handle.data = state;
-
-        int initResult = uv_poll_init_socket(runtime->getEventLoop(), &state->handle, socket);
-        if (initResult != 0)
-        {
-            delete state;
-            return false;
-        }
-
-        pollState = state;
-
-        if (!updatePollingInterest())
-        {
-            stopPolling();
-            return false;
-        }
-
-        return true;
-    }
-
-    void flushOutgoing()
-    {
-        while (!pendingSends.empty() && !isClosed.load())
-        {
-            PendingSend& pending = pendingSends.front();
-
-            size_t sent = 0;
-            CURLcode result = CURLE_OK;
-
-            {
-                std::lock_guard<std::mutex> lock(curlMutex);
-                if (!curl)
-                {
-                    clearPendingSends();
-                    return;
-                }
-
-                size_t remaining = pending.payload.size() - pending.offset;
-                const char* data = remaining > 0 ? pending.payload.data() + pending.offset : pending.payload.data();
-
-                result =
-                    curl_ws_send(curl, data, remaining, &sent, 0, pending.binary ? CURLWS_BINARY : CURLWS_TEXT);
-            }
-
-            if (result == CURLE_AGAIN)
-                break;
-
-            if (result != CURLE_OK)
-            {
-                closeWithError("websocket send failed: " + std::string(curl_easy_strerror(result)));
-                return;
-            }
-
-            pending.offset += sent;
-
-            if (pending.payload.empty() || pending.offset >= pending.payload.size())
-            {
-                pendingSends.pop_front();
-                continue;
-            }
-
-            if (sent == 0)
-                break;
-        }
-
-        if (!isClosed.load() && !updatePollingInterest())
-            closeWithError("failed to update websocket polling");
-    }
-
-    void close()
-    {
-        closeWithCode();
-    }
-
-    ~WebSocketHandle()
-    {
-        close();
-
-        onOpenRef.reset();
-        onMessageRef.reset();
-        onCloseRef.reset();
-        onErrorRef.reset();
-    }
-};
-
-static std::shared_ptr<WebSocketHandle>* getWebSocketHandle(lua_State* L, int index)
-{
-    return static_cast<std::shared_ptr<WebSocketHandle>*>(lua_touserdatatagged(L, index, kWebSocketHandleTag));
-}
-
 int request(lua_State* L)
 {
     std::string url = luaL_checkstring(L, 1);
@@ -726,192 +288,6 @@ int request(lua_State* L)
 
     return lua_yield(L, 0);
 }
-
-int ws_send(lua_State* L)
-{
-    if (lua_gettop(L) != 2)
-        luaL_errorL(L, "websocket send expects exactly 1 payload argument");
-
-    luaL_checktype(L, 1, LUA_TUSERDATA);
-    auto* handleStorage = getWebSocketHandle(L, 1);
-    if (!handleStorage || !(*handleStorage) || (*handleStorage)->isClosed.load())
-        luaL_errorL(L, "Invalid or closed websocket");
-
-    std::shared_ptr<WebSocketHandle> handle = *handleStorage;
-
-    WebSocketPayload payloadData = extractWebSocketPayload(L, 2);
-    handle->pendingSends.push_back({std::string(payloadData.data, payloadData.length), payloadData.binary, 0});
-    handle->flushOutgoing();
-    return 0;
-}
-
-int ws_close(lua_State* L)
-{
-    luaL_checktype(L, 1, LUA_TUSERDATA);
-    auto* handleStorage = getWebSocketHandle(L, 1);
-    if (!handleStorage || !(*handleStorage))
-        luaL_errorL(L, "Invalid websocket");
-
-    (*handleStorage)->close();
-    return 0;
-}
-
-int websocket(lua_State* L)
-{
-    std::string url = luaL_checkstring(L, 1);
-    std::vector<std::pair<std::string, std::string>> headers;
-    std::shared_ptr<Ref> onOpenRef;
-    std::shared_ptr<Ref> onMessageRef;
-    std::shared_ptr<Ref> onCloseRef;
-    std::shared_ptr<Ref> onErrorRef;
-
-    if (lua_istable(L, 2))
-    {
-        lua_getfield(L, 2, "headers");
-        if (lua_istable(L, -1))
-        {
-            lua_pushnil(L);
-            while (lua_next(L, -2))
-            {
-                if (lua_isstring(L, -2) && lua_isstring(L, -1))
-                {
-                    std::string key = lua_tostring(L, -2);
-                    std::string value = lua_tostring(L, -1);
-                    headers.emplace_back(key, value);
-                }
-                lua_pop(L, 1);
-            }
-        }
-        lua_pop(L, 1);
-
-        lua_getfield(L, 2, "onopen");
-        if (lua_isfunction(L, -1))
-            onOpenRef = std::make_shared<Ref>(L, -1);
-        lua_pop(L, 1);
-
-        lua_getfield(L, 2, "onmessage");
-        if (lua_isfunction(L, -1))
-            onMessageRef = std::make_shared<Ref>(L, -1);
-        lua_pop(L, 1);
-
-        lua_getfield(L, 2, "onclose");
-        if (lua_isfunction(L, -1))
-            onCloseRef = std::make_shared<Ref>(L, -1);
-        lua_pop(L, 1);
-
-        lua_getfield(L, 2, "onerror");
-        if (lua_isfunction(L, -1))
-            onErrorRef = std::make_shared<Ref>(L, -1);
-        lua_pop(L, 1);
-    }
-
-    auto token = getResumeToken(L);
-    Runtime* runtime = token->runtime;
-
-    token->runtime->runInWorkQueue(
-        [token,
-         runtime,
-         url = std::move(url),
-         headers = std::move(headers),
-         onOpenRef = std::move(onOpenRef),
-         onMessageRef = std::move(onMessageRef),
-         onCloseRef = std::move(onCloseRef),
-         onErrorRef = std::move(onErrorRef)]() mutable
-        {
-            CURL* curl = curl_easy_init();
-            if (!curl)
-            {
-                token->fail("failed to initialize websocket");
-                return;
-            }
-
-            curl_slist* headerList = nullptr;
-
-            curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
-            curl_easy_setopt(curl, CURLOPT_SSL_OPTIONS, CURLSSLOPT_NATIVE_CA);
-            curl_easy_setopt(curl, CURLOPT_CONNECT_ONLY, 2L);
-
-            for (const auto& headerPair : headers)
-            {
-                std::string headerString = headerPair.first + ": " + headerPair.second;
-                headerList = curl_slist_append(headerList, headerString.c_str());
-            }
-
-            if (headerList)
-                curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headerList);
-
-            CURLcode result = curl_easy_perform(curl);
-            if (result != CURLE_OK)
-            {
-                std::string error = curl_easy_strerror(result);
-                if (headerList)
-                    curl_slist_free_all(headerList);
-                curl_easy_cleanup(curl);
-                token->fail("websocket connect failed: " + error);
-                return;
-            }
-
-            curl_socket_t socket = CURL_SOCKET_BAD;
-            if (curl_easy_getinfo(curl, CURLINFO_ACTIVESOCKET, &socket) != CURLE_OK || socket == CURL_SOCKET_BAD)
-            {
-                if (headerList)
-                    curl_slist_free_all(headerList);
-                curl_easy_cleanup(curl);
-                token->fail("failed to get websocket socket");
-                return;
-            }
-
-            token->complete(
-                [runtime,
-                 curl,
-                 socket,
-                 headerList,
-                 onOpenRef = std::move(onOpenRef),
-                 onMessageRef = std::move(onMessageRef),
-                 onCloseRef = std::move(onCloseRef),
-                 onErrorRef = std::move(onErrorRef)](lua_State* L) mutable
-                {
-                    auto handle = std::make_shared<WebSocketHandle>();
-                    handle->runtime = runtime;
-                    handle->curl = curl;
-                    handle->headerList = headerList;
-                    handle->onOpenRef = std::move(onOpenRef);
-                    handle->onMessageRef = std::move(onMessageRef);
-                    handle->onCloseRef = std::move(onCloseRef);
-                    handle->onErrorRef = std::move(onErrorRef);
-
-                    if (!handle->startPolling(socket))
-                    {
-                        handle->finishClose(false);
-                        luaL_errorL(L, "failed to initialize websocket polling");
-                    }
-
-                    auto* storage =
-                        new (static_cast<std::shared_ptr<WebSocketHandle>*>(lua_newuserdatataggedwithmetatable(
-                            L,
-                            sizeof(std::shared_ptr<WebSocketHandle>),
-                            kWebSocketHandleTag
-                        ))) std::shared_ptr<WebSocketHandle>(handle);
-                    (void)storage;
-
-                    runtime->addPendingToken();
-                    handle->hasPendingToken.store(true);
-                    handle->scheduleCallback(
-                        handle->onOpenRef,
-                        [](lua_State*)
-                        {
-                            return 0;
-                        }
-                    );
-                    return 1;
-                }
-            );
-        }
-    );
-
-    return lua_yield(L, 0);
-}
-
 } // namespace net::client
 
 const char* const NetClient::properties[] = {nullptr};
